turns out city-state needs read/write gem perms
it extracts data on demand from the maxmind database, but that means
that (for example) when we try to list states for the agency signup
form, it blows up if it can't write out the extracted list of states
into the filesystem. which it can't, if the gem is owned by root.