simplify docker tiers
The base image creates a snapshot of all dependencies, making it
very fast to update bundler/yarn deps during regular builds.

At this moment, regular builds take about 65 seconds, which is excellent
and exactly where I would like to keep them for the forseeable future.

I think we'll want to update the base image maybe once every few days,
so that the on-push builds stay as fast as possible.