ughhhh engines do weird path things
so if Passwordless is mounted at /survey, it somehow inserts that at the
front of _all_ paths, even paths that are unrelated. For example, if you
are on the Passwordless login page, suddenly the navbar
/agency/dashboard becomes /survey/agency/dashboard instead.

The only workaround that seems immediately doable after some initial
googling is to just move Passwordless to the root, so that it doesn't
have anything to prepend to those routes and they will continue to work
correctly.